Font family and size components moved to different files
