+++ package.json
@@ -20,7 +20,6 @@
     "@tiptap/react": "^2.9.1",
     "@tiptap/starter-kit": "^2.9.1",
     "axios": "^1.7.7",
-    "axois": "^0.0.1-security",
     "bcrypt": "^5.1.1",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
